Rename and move mlflow charm
Renames mlflow to mlflow-server, and moves to a charms/ subdirectory

Also updates it to use consistent style with other charms
