+++ bundle.yaml
@@ -1,33 +1,29 @@
 ---
 applications:
-  mlflow:
-    charm: "./mlflow.charm"
+  mlflow-server:
+    charm: "cs:mlflow-server"
     scale: 1
 
   mysql:
-    charm: "cs:~charmed-osm/mariadb-k8s-34"
+    charm: "cs:~charmed-osm/mariadb-k8s-35"
     scale: 1
 
-  # avoid colliding with name of minio for kubeflow, which uses a newer charm
-  # which exposes an 'object-storage' interface rather than a 'minio'
-  # interface, which the mlflow charm can't speak yet.
-  minio-old:
-    charm: "cs:~kubeflow-charmers/minio-146"
+  minio:
+    charm: "cs:minio"
     scale: 1
 
   prometheus:
     charm: "cs:~charmed-osm/prometheus-k8s-32"
     options:
-      default-target: 'mlflow:5000'
+      default-target: 'mlflow-server:5000'
     scale: 1
 
 bundle: kubernetes
 description: "MLflow"
 relations:
-  - - mlflow:db
+  - - mlflow-server
     - mysql
-  - - mlflow:minio
-    - minio-old
-    #- minio:object-storage
-  - - mlflow:prometheus
+  - - mlflow-server:object-storage
+    - minio:object-storage
+  - - mlflow-server
     - prometheus
