+++ charms/mlflow-server/metadata.yaml
@@ -1,0 +1,30 @@
+# Copyright 2020 Luke Marsden
+# See LICENSE file for licensing details.
+name: mlflow-server
+description: |
+  MLflow
+summary: |
+  A charm which provides a Kubernetes installation of the MLflow (https://mlflow.org/) server.
+series: [kubernetes]
+tags: [mlflow]
+resources:
+  oci-image:
+    type: oci-image
+    description: Backing OCI image
+    auto-fetch: true
+    upstream-source: quay.io/helix-ml/mlflow:1.13.1
+requires:
+  db:
+    interface: mysql
+  object-storage:
+    interface: object-storage
+    schema: https://raw.githubusercontent.com/canonical/operator-schemas/master/object-storage.yaml
+    versions: [v1]
+  ingress:
+    interface: ingress
+    schema: https://raw.githubusercontent.com/canonical/operator-schemas/master/ingress.yaml
+    versions: [v1]
+  prometheus:
+    interface: prometheus
+  pod-defaults:
+    interface: pod-defaults
