+++ charms/mlflow-server/test-requirements.txt
@@ -1,0 +1,3 @@
+black
+flake8
+pytest
