+++ charms/mlflow-server/tests/test_charm.py
@@ -1,0 +1,44 @@
+import pytest
+
+from ops.model import BlockedStatus, WaitingStatus
+from ops.testing import Harness
+import yaml
+
+from charm import Operator
+
+
+@pytest.fixture
+def harness():
+    return Harness(Operator)
+
+
+def test_not_leader(harness):
+    harness.begin()
+    assert isinstance(harness.charm.model.unit.status, WaitingStatus)
+
+
+def test_missing_image(harness):
+    harness.set_leader(True)
+    harness.begin_with_initial_hooks()
+    assert isinstance(harness.charm.model.unit.status, BlockedStatus)
+
+
+def test_main_no_relation(harness):
+    harness.set_leader(True)
+    harness.add_oci_resource(
+        "oci-image",
+        {
+            "registrypath": "ci-test",
+            "username": "",
+            "password": "",
+        },
+    )
+    harness.begin_with_initial_hooks()
+    pod_spec = harness.get_pod_spec()
+
+    # confirm that we can serialize the pod spec
+    yaml.safe_dump(pod_spec)
+
+    assert harness.charm.model.unit.status == WaitingStatus(
+        "Waiting for mysql relation data"
+    )
