+++ charms/mlflow-server/tox.ini
@@ -1,0 +1,21 @@
+[flake8]
+max-line-length = 100
+
+[tox]
+skipsdist = True
+
+[testenv]
+setenv =
+	PYTHONPATH={toxinidir}/src:{toxinidir}/lib
+deps =
+	-rtest-requirements.txt
+	-rrequirements.txt
+
+[testenv:unit]
+commands = 
+	pytest tests
+
+[testenv:lint]
+commands =
+    flake8 {toxinidir}/src {toxinidir}/tests
+    black --check {toxinidir}/src {toxinidir}/tests
