--- mlflow/metadata.yaml
@@ -1,26 +1,0 @@
-# Copyright 2020 Luke Marsden
-# See LICENSE file for licensing details.
-name: mlflow
-description: |
-  MLflow
-summary: |
-  A charm which provides a Kubernetes installation of the MLflow (https://mlflow.org/) server.
-series: [kubernetes]
-tags: [mlflow]
-#resources:
-#  oci-image:
-#    type: oci-image
-#    description: Backing OCI image
-#    auto-fetch: true
-requires:
-  db:
-    interface: mysql
-    limit: 1  # Most charms only handle a single MySQL Application.
-  minio:
-    interface: generic-ip-port-user-pass
-    limit: 1
-  prometheus:
-    interface: prometheus
-    limit: 1
-  pod-defaults:
-    interface: pod-defaults
