--- mlflow/requirements-dev.txt
@@ -1,2 +1,0 @@
--r requirements.txt
-flake8
