--- mlflow/requirements.txt
@@ -1,5 +1,0 @@
-ops==0.8.0
-git+git://github.com/johnsca/resource-oci-image.git#oci_image
-ops-lib-mysql
-minio
-charmhelpers
